backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.30.0
psycopg2-binary==2.9.9
redis==5.1.1
celery==5.4.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
anthropic==0.40.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
tenacity==9.0.0
fake-useragent==1.5.1
stripe==11.0.0
email-validator==2.2.0
